refactor(user): move user classes into user package
