+++ tracker-api/src/main/java/com/sloyardms/trackerapi/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.sloyardms.trackerapi.repository;
+package com.sloyardms.trackerapi.user;
 
 import com.sloyardms.trackerapi.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
